Rename and restructure to avoid confusion
Server-side
* Controller: Primary expedite server that launches the agents
* Agent: The user application that is used as the clean source for forking

Client-side
Both commands will fork from an agent
* Exec: Runs an action in the specified agent and replace the current executable
* Invoke: Runs an action in the specified agent and returns the value

Updated public interface to use `Expedite.define`

Cleaned up output of examples
