+++ examples/rails/README.md
@@ -1,24 +1,5 @@
-# README
+Run the script/list_user.rb in the worker process
 
-This README would normally document whatever steps are necessary to get the
-application up and running.
-
-Things you may want to cover:
-
-* Ruby version
-
-* System dependencies
-
-* Configuration
-
-* Database creation
-
-* Database initialization
-
-* How to run the test suite
-
-* Services (job queues, cache servers, search engines, etc.)
-
-* Deployment instructions
-
-* ...
+```
+$ bundle exec ruby script/run.rb script/list_user.rb
+```
