+++ examples/rails/script/run.rb
@@ -1,16 +1,18 @@
 #!/usr/bin/env ruby
 require 'expedite'
 
-ret = Expedite.v("development/abc").invoke("runner", ARGV)
-puts "ret = #{ret}"
+label = "[#{Process.pid}] run.rb"
+
+ret = Expedite.agent("development/abc").invoke("runner", ARGV)
+puts "#{label}: Invoked runner on development/abc, got #{ret}"
 
 begin
-  ret = Expedite.v("development/abc").invoke("raise")
-  puts "ret = #{ret}"
+  ret = Expedite.agent("development/abc").invoke("raise")
+  puts "#{label}: We never reach here"
 rescue => e
-  puts "#{e}: #{e.backtrace.join("\n")}"
+  puts "#{label}: Invoked raise on development/abc, caught: #{e}\n  #{e.backtrace.join("\n  ")}"
 end
 
-ret = Expedite.v("development/abc").exec("runner", ARGV)
-# never reaches here
-puts "ret = #{ret}"
+puts "#{label}: Exec runner on development/abc. Process will be replaced..."
+Expedite.agent("development/abc").exec("runner", ARGV)
+puts "#{label}: We never reach here"
