+++ examples/simple/expedite_helper.rb
@@ -1,19 +1,18 @@
-require 'expedite/commands'
-require 'expedite/variants'
+require 'expedite'
 
-# You can pass `keep_alive: true` if you want the variant to restart
-# automatically if it is terminated. This option defaults to false.
-Expedite::Variants.register('parent') do
-  $sleep_parent = 1
-end
+Expedite.define do
+  agent :parent do
+    $parent_var = 1
+  end
 
-Expedite::Variants.register('development/*', parent: 'parent') do |name|
-  $sleep_child = name
-end
+  agent "development/*", parent: :parent do |name|
+    $development_var = name
+  end
 
-Expedite::Commands.register("custom") do
-  puts "[#{Expedite.variant}] sleeping for 5"
-  puts "$sleep_parent = #{$sleep_parent}"
-  puts "$sleep_child = #{$sleep_child}"
-  puts "[#{Expedite.variant}] done"
+  action :info do
+    puts "     Process.pid = #{Process.pid}"
+    puts "    Process.ppid = #{Process.ppid}"
+    puts "     $parent_var = #{$parent_var}"
+    puts "$development_var = #{$development_var}"
+  end
 end
