+++ examples/simple/main.rb
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 require 'expedite'
 
-Expedite.v("development/abc").call("custom")
+Expedite.agent("development/abc").invoke("info")
