+++ lib/expedite/actions.rb
@@ -1,10 +1,10 @@
-require 'expedite/command/basic'
-require 'expedite/command/boot'
+require 'expedite/action/block'
+require 'expedite/action/boot'
 
 module Expedite
-  class Commands
+  class Actions
     def self.current
-      @current ||= Commands.new
+      @current ||= Actions.new
     end
 
     def self.lookup(name)
@@ -12,16 +12,16 @@ module Expedite
     end
 
     ##
-    # Registers a command. If multiple commands are registered with the
+    # Registers an action. If multiple actions are registered with the
     # same name, the last one takes precedence.
     #
-    # [name] Name of the command. Expedite internal commands are prefixed
+    # [name] Name of the action. Expedite internal actions are prefixed
     #        with "expedite/"
-    # [klass_or_nil] Class of the command. If omitted, will default to
-    #                Expedite::Command::Basic.
-    # [named_options] Command options. Passed to the initializer.
+    # [klass_or_nil] Class of the action. If omitted, will default to
+    #                Expedite::Action::Block.
+    # [named_options] Action options. Passed to the initializer.
     def self.register(name, klass_or_nil = nil, **named_options, &block)
-      self.current.register(name, klass_or_nil, **named_options, &block)
+      self.current.register(name.to_s, klass_or_nil, **named_options, &block)
     end
 
     ##
@@ -36,13 +36,13 @@ module Expedite
 
     def lookup(name)
        ret = @registrations[name]
-       raise NotImplementedError, "Command #{name.inspect} not found" if ret.nil?
+       raise NotImplementedError, "Action #{name.inspect} not found" if ret.nil?
        ret
     end
 
     def register(name, klass_or_nil = nil, **named_options, &block)
       cmd = if klass_or_nil.nil?
-        Command::Basic.new(**named_options, &block)
+        Action::Block.new(**named_options, &block)
       else
         klass_or_nil.new(**named_options)
       end
@@ -54,7 +54,7 @@ module Expedite
       @registrations = {}
 
       # Default registrations
-      register("expedite/boot", Expedite::Command::Boot)
+      register("expedite/boot", Expedite::Action::Boot)
 
       nil
     end
