--- lib/expedite/variants.rb
@@ -1,100 +1,0 @@
-
-module Expedite
-  # Definition of a Variant
-  class Variant
-    ##
-    # Name of the parent variant. This allows you to create variants from
-    # an existing variant.
-    # Defaults to nil.
-    attr_accessor :parent
-
-    ##
-    # If set to true, variant will be restarted automatically if it is killed.
-    # Defaults to false.
-    attr_accessor :keep_alive
-
-    ##
-    # [parent] Name of parent variant.
-    # [after_fork] Block is executed when variant is first preloaded.
-    def initialize(parent: nil, keep_alive: false, &after_fork)
-      @parent = parent
-      @keep_alive = keep_alive
-      @after_fork_proc = after_fork
-    end
-
-    ##
-    # Called when variant if first preloaded. This version calls the after_fork
-    # block provided in the initializer.
-    def after_fork(variant)
-      @after_fork_proc&.call(variant)
-    end
-  end
-
-  class Variants
-    Registration = Struct.new(:matcher, :variant) do
-      def match?(name)
-        File.fnmatch?(matcher, name)
-      end
-    end
-
-    def self.current
-      @current ||= Variants.new
-    end
-
-    ##
-    # Retrieves the specified variant
-    def self.lookup(variant)
-      self.current.lookup(variant)
-    end
-
-    ##
-    # Registers a variant. Variants are matched in the
-    # order they are registered.
-    #
-    # [matcher] Wildcard to match a name against.
-    # [named_options] Variant options.
-    # [after_fork] Optional block that is called when
-    #              variant is preloaded.
-    #
-    # = Example
-    #   Expedite::Variants.register('base' do |name|
-    #     puts "Base #{name} started"
-    #   end
-    #   Expedite::Variants.register('development/abc', parent: 'base') do |name|
-    #     puts "Variant #{name} started"
-    #   end
-    def self.register(matcher, **named_options, &after_fork)
-      self.current.register(matcher, **named_options, &after_fork)
-    end
-
-    ##
-    # Resets registrations to default
-    def self.reset
-      self.current.reset
-    end
-
-    def initialize
-      @registrations = []
-    end
-
-    def lookup(variant)
-      ret = @registrations.find do |r|
-        r.match?(variant)
-      end
-      raise NotImplementedError, "Variant #{variant.inspect} not found" if ret.nil?
-      ret.variant
-    end
-
-    def register(matcher, **named_options, &after_fork)
-      @registrations << Registration.new(
-        matcher,
-        Variant.new(**named_options, &after_fork)
-      )
-    end
-
-    def reset
-      @registrations = []
-      nil
-    end
-  end
-end
