+++ test/commands_test.rb
@@ -1,27 +1,27 @@
 require "test_helper"
-require "expedite/commands"
+require "expedite/actions"
 
 class CommandsTest < Minitest::Test
   def test_register
     assert_raises ::NotImplementedError do
-      Expedite::Commands.lookup("missing")
+      Expedite::Actions.lookup("missing")
     end
 
-    Expedite::Commands.register("dev") do
+    Expedite::Actions.register("dev") do
       10
     end
 
-    cmd = Expedite::Commands.lookup("dev")
+    cmd = Expedite::Actions.lookup("dev")
     assert cmd != nil
     assert cmd.call == 10
 
-    Expedite::Commands.reset
+    Expedite::Actions.reset
     assert_raises ::NotImplementedError do
-      Expedite::Commands.lookup("dev")
+      Expedite::Actions.lookup("dev")
     end
   end
 
   def test_boot
-    assert Expedite::Commands.lookup("expedite/boot") != nil
+    assert Expedite::Actions.lookup("expedite/boot") != nil
   end
 end
