+++ test/fixtures/expedite_helper.rb
@@ -1,4 +1,4 @@
 # Registrations used in ApplicationManagerTest
-Expedite::Variants.register("parent")
-Expedite::Variants.register("test", parent: "parent", keep_alive: true) do
+Expedite::Agents.register("parent")
+Expedite::Agents.register("test", parent: "parent", keep_alive: true) do
 end
